Changes in stacks
- aws: added calls in rpc.go
- aws: added handling of aws.Error with code "InvalidParameterValue"
- aws: changes Security Group rule convert to AWS counterpart to support use of Security Group ID as source/target, supplementary to IP ranges (but no mix allowed and checked)
- all: Security Group rule port type replaced from uint16 to int32 (may contain -1 meaning all ports)
- all: consolidated abstract.HostFull.CurrentState values
+ code cleanup

+++ lib/server/iaas/stacks/outscale/securitygroup.go
@@ -72,8 +72,8 @@ func toAbstractSecurityGroupRule(in osc.SecurityGroupRule, direction securitygro
 	out := abstract.SecurityGroupRule{
 		Direction: direction,
 		Protocol:  in.IpProtocol,
-		PortFrom:  uint16(in.FromPortRange),
-		PortTo:    uint16(in.ToPortRange),
+		PortFrom:  int32(in.FromPortRange),
+		PortTo:    int32(in.ToPortRange),
 		IPRanges:  in.IpRanges,
 	}
 	return out
